DRTII-715 Move existing deployment tests to more appropriate location
